Add ondelete clauses for other fields on column. Rename connection to base.
