+++ datajunction-server/datajunction_server/database/database.py
@@ -8,8 +8,8 @@ from sqlalchemy import JSON, BigInteger, DateTime, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy_utils import UUIDType
 
+from datajunction_server.database.base import Base
 from datajunction_server.database.column import Column
-from datajunction_server.database.connection import Base
 from datajunction_server.typing import UTCDatetime
