+++ datajunction-server/datajunction_server/database/engine.py
@@ -4,7 +4,7 @@ from typing import Optional
 import sqlalchemy as sa
 from sqlalchemy.orm import Mapped, mapped_column
 
-from datajunction_server.database.connection import Base
+from datajunction_server.database.base import Base
 from datajunction_server.models.base import sqlalchemy_enum_with_name
 from datajunction_server.models.engine import Dialect
