+++ datajunction-server/datajunction_server/database/measure.py
@@ -4,8 +4,8 @@ from typing import List, Optional
 from sqlalchemy import BigInteger, Enum, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
+from datajunction_server.database.base import Base
 from datajunction_server.database.column import Column
-from datajunction_server.database.connection import Base
 from datajunction_server.models.base import labelize
 from datajunction_server.models.measure import AggregationRule
