+++ datajunction-server/datajunction_server/database/user.py
@@ -4,7 +4,7 @@ from typing import Optional
 from sqlalchemy import BigInteger, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
-from datajunction_server.database.connection import Base
+from datajunction_server.database.base import Base
 from datajunction_server.enum import StrEnum
 from datajunction_server.models.base import sqlalchemy_enum_with_name
